+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * Added prop to disable mouse wheel action.
+* Added prop to add title attribute on the wheel.
 
 # 0.3.0
 * Added boolean property `disableTextInput`